+++ CHANGELOG.md
@@ -10,7 +10,7 @@ Alternatively view it [online on GitHub](https://github.com/cumulusmx/CumulusMX/
 ---
 ---
 
-## [4.6.0 \[b4107\]][23] - 2025-08-07
+## [4.6.0 \[b4108\]][23] - 2025-08-10
 
 ### New
 
@@ -956,4 +956,4 @@ Initial release of Cumulus MX which now runs under Microsoft .NET 8.0 and remove
 [20]: https://github.com/cumulusmx/CumulusMX/releases/tag/b4103
 [21]: https://github.com/cumulusmx/CumulusMX/releases/tag/b4104
 [22]: https://github.com/cumulusmx/CumulusMX/releases/tag/b4105
-[23]: https://github.com/cumulusmx/CumulusMX/releases/tag/b4107
+[23]: https://github.com/cumulusmx/CumulusMX/releases/tag/b4108